Interface: 메인 api 인터페이스 정의
