+++ src/models/sports.model.ts
@@ -1,0 +1,5 @@
+export interface ISport {
+  sportId: number;
+  sportName: string;
+  sportImageUrl: string;
+}
